requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
astroid==2.15.4
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.2
bing-image-downloader==1.1.2
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
cryptography==39.0.2
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
dj-database-url==1.2.0
Django==4.1.7
django-allauth==0.53.1
django-ckeditor==6.5.1
django-countries==7.5.1
django-environ==0.10.0
django-filter==23.2
django-js-asset==2.0.0
django-social-share==2.3.0
django-taggit==4.0.0
djangorestframework==3.14.0
downloader==0.98
et-xmlfile==1.1.0
exceptiongroup==1.1.1
filelock==3.11.0
frozenlist==1.3.3
google-images-download==2.8.0
greenlet==2.0.1
gunicorn==20.1.0
h11==0.14.0
idna==3.4
image==1.5.33
isort==5.12.0
lazy-object-proxy==1.9.0
Markdown==3.4.3
mccabe==0.7.0
multidict==6.0.4
numpy==1.24.2
oauthlib==3.2.2
openai==0.27.3
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.0.0
pexels-api==1.0.1
Pillow==9.4.0
platformdirs==3.2.0
playwright==1.32.1
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pyee==9.0.4
PyJWT==2.6.0
pylint==2.17.4
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-stripe==0.0.5
python3-openid==3.2.0
pytz==2021.3
requests==2.28.2
requests-oauthlib==1.3.1
selenium==4.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.4.3
stripe==2.66.0
tomli==2.0.1
tomlkit==0.11.8
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.1.1
tzdata==2023.3
urllib3==1.26.15
virtualenv==20.21.0
webdriver-manager==3.8.5
whitenoise==5.3.0
wrapt==1.15.0
wsproto==1.2.0
yarl==1.9.2
